// browser-extension/src/client/generated/models/TutorialBrief.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { TutorialBriefStepsInner } from './TutorialBriefStepsInner';
import {
    TutorialBriefStepsInnerFromJSON,
    TutorialBriefStepsInnerFromJSONTyped,
    TutorialBriefStepsInnerToJSON,
} from './TutorialBriefStepsInner';

/**
 * 
 * @export
 * @interface TutorialBrief
 */
export interface TutorialBrief {
    /**
     * 
     * @type {number}
     * @memberof TutorialBrief
     */
    readonly id?: number;
    /**
     * 
     * @type {string}
     * @memberof TutorialBrief
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof TutorialBrief
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof TutorialBrief
     */
    targetSite: string;
    /**
     * 
     * @type {Array<TutorialBriefStepsInner>}
     * @memberof TutorialBrief
     */
    readonly steps?: Array<TutorialBriefStepsInner>;
}

/**
 * Check if a given object implements the TutorialBrief interface.
 */
export function instanceOfTutorialBrief(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "title" in value;
    isInstance = isInstance && "targetSite" in value;

    return isInstance;
}

export function TutorialBriefFromJSON(json: any): TutorialBrief {
    return TutorialBriefFromJSONTyped(json, false);
}

export function TutorialBriefFromJSONTyped(json: any, ignoreDiscriminator: boolean): TutorialBrief {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'title': json['title'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'targetSite': json['target_site'],
        'steps': !exists(json, 'steps') ? undefined : ((json['steps'] as Array<any>).map(TutorialBriefStepsInnerFromJSON)),
    };
}

export function TutorialBriefToJSON(value?: TutorialBrief | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
        'description': value.description,
        'target_site': value.targetSite,
    };
}